requirements.txt
pytorch-lightning==1.2.5
https://download.pytorch.org/whl/cpu/torch-1.8.1%2Bcpu-cp39-cp39-linux_x86_64.whl
nltk==3.4.5
hydra-core==1.1.0.dev5
wandb==0.12.6
transformers==4.12.3
torchtext==0.9.1
black==21.9b0
python-dotenv==0.19.1
streamlit==1.3.0
spacy==3.2.1
htbuilder==0.5.0
gdown==4.2.0
